Transition enums to string constants. Fixes a couple other minor ESLint issues.
I'm not entirely sure about the benefits of this transition, but it presents us a couple nice-to-haves:

(1) It is now much easier to share types between the worker and the frontend. This is non-const enums compile down to a JavaScript object that cannot be purely removed at compile time. Previously, incident types could not be fully shared between the type, because it would require a much more intricate bundler set up.

(2) Enables us to use fast-refresh more easily, which massively increases developer productivity

(3) Using const enums would likely get a lot of the benefits of (1) and (2) but they can be unpredictable to developers, and are generally frowned upon and a not well-supported part of TypeScript

The downsides of this move are several:

(1) Makes refactors in the future harder, because consumers are now using strings instead of accessing known properties (ButtonMode.Normal becomes "normal"). This can be mitigated by careful find-and-replace, but would slow developer velocity when making these refactors.

(2) Very slight performance degradation, as the underlying type of enums (normally number) is easier to compare than strings. I don't anticipate this to be a measurable difference for a project of this size

(3) It is harder to document what specific variants in string unions mean, whereas enums can get comments with explanations.

Ultimately, I am convinced that this is currently the right thing to do at this point, before the project expands too much in scope (or gains too many other contributors). The benefits of fast refresh lead to a significant increase in developer productivity.