Use bulk operations to make server_share_info handling MUCH more efficient.
A lot of the previous delay and issues we have seen is due to crazy I/O bloat with how this was being handled for larger events. Streamlining IndexedDB access into fewer transactions and acting on things in bulk can massively improve performance.

This change should be a nearly 50x speed up for larger events. In my experimentation, an event with 2000 incidents took around 15 seconds if it was empty locally, and now should take around 200ms