Don't pass undefined to RulesMultiSelect.
Now the behavior if a rule does not have an incident is to silently not display it in the list when you open for editing. However, it will display in the incident view.

I think this behavior is ok, assuming that we are not deploying this change the in the middle of an event.